Highlight breakpoints in `nearpc` output (#2828)
* Expose list of breakpoint locations in aglib

The list of addresses with active breakpoints can be retrieved using `Debugger.breakpoint_locations()`.

* Highlight breakpoints in `nearpc` output

Addresses in the disassembly which have an active breakpoint attached to them
are prefixed using by `b+` and highlighted in red by default.

This can be configured using the new `highlight-breakpoints`, `nearpc-breakpoint`,
and `nearpc-breakpoint-color` theme config options.

* Fix `breakpoint_locations` on lldb

* Update tests/gdb-tests/tests/test_nearpc.py

* Refactor breakpoint prefix logic

* Account for leading space before prefix again

* Move prefix coloring and preparation outside loop

---------

Co-authored-by: Disconnect3d <dominik.b.czarnota@gmail.com>