TUI: Switch threads and expressions windows (#3095)
The order of the context section windows in the  pre-configured pwndbg layouts didn't match the order in which they are drawn in the CLI context output as specified by the `context-sections` config. This caused glitches while rendering the section titles.

This is a workaround for the current TUI layout and the context-sections order. If the order is changed later, it might break again.

Fixes #3045